refactor: added new csprojects moved corresponding files in their new directories now fixing usings and namepsaces
