+++ CtfDeck.sln
@@ -11,6 +11,12 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{0AB3BF05
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CtfDeck.Tests", "tests\CtfDeck.Tests\CtfDeck.Tests.csproj", "{A8EDFA8B-D4CB-4C33-90D6-40E45849C773}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CtfDeck.Contracts", "src\CtfDeck.Contracts\CtfDeck.Contracts.csproj", "{DB055AC4-1685-4A24-9779-7E1A37CA9986}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CtfDeck.Data", "src\CtfDeck.Data\CtfDeck.Data.csproj", "{BA68D2AA-B0BE-4C8B-AD7C-B374A716865F}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CtfDeck.ServerWs", "src\CtfDeck.ServerWs\CtfDeck.ServerWs.csproj", "{4622F145-00FA-4BE3-939C-2CD67CAFC958}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -45,6 +51,42 @@ Global
 		{A8EDFA8B-D4CB-4C33-90D6-40E45849C773}.Release|x64.Build.0 = Release|Any CPU
 		{A8EDFA8B-D4CB-4C33-90D6-40E45849C773}.Release|x86.ActiveCfg = Release|Any CPU
 		{A8EDFA8B-D4CB-4C33-90D6-40E45849C773}.Release|x86.Build.0 = Release|Any CPU
+		{DB055AC4-1685-4A24-9779-7E1A37CA9986}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{DB055AC4-1685-4A24-9779-7E1A37CA9986}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{DB055AC4-1685-4A24-9779-7E1A37CA9986}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{DB055AC4-1685-4A24-9779-7E1A37CA9986}.Debug|x64.Build.0 = Debug|Any CPU
+		{DB055AC4-1685-4A24-9779-7E1A37CA9986}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{DB055AC4-1685-4A24-9779-7E1A37CA9986}.Debug|x86.Build.0 = Debug|Any CPU
+		{DB055AC4-1685-4A24-9779-7E1A37CA9986}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{DB055AC4-1685-4A24-9779-7E1A37CA9986}.Release|Any CPU.Build.0 = Release|Any CPU
+		{DB055AC4-1685-4A24-9779-7E1A37CA9986}.Release|x64.ActiveCfg = Release|Any CPU
+		{DB055AC4-1685-4A24-9779-7E1A37CA9986}.Release|x64.Build.0 = Release|Any CPU
+		{DB055AC4-1685-4A24-9779-7E1A37CA9986}.Release|x86.ActiveCfg = Release|Any CPU
+		{DB055AC4-1685-4A24-9779-7E1A37CA9986}.Release|x86.Build.0 = Release|Any CPU
+		{BA68D2AA-B0BE-4C8B-AD7C-B374A716865F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{BA68D2AA-B0BE-4C8B-AD7C-B374A716865F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{BA68D2AA-B0BE-4C8B-AD7C-B374A716865F}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{BA68D2AA-B0BE-4C8B-AD7C-B374A716865F}.Debug|x64.Build.0 = Debug|Any CPU
+		{BA68D2AA-B0BE-4C8B-AD7C-B374A716865F}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{BA68D2AA-B0BE-4C8B-AD7C-B374A716865F}.Debug|x86.Build.0 = Debug|Any CPU
+		{BA68D2AA-B0BE-4C8B-AD7C-B374A716865F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{BA68D2AA-B0BE-4C8B-AD7C-B374A716865F}.Release|Any CPU.Build.0 = Release|Any CPU
+		{BA68D2AA-B0BE-4C8B-AD7C-B374A716865F}.Release|x64.ActiveCfg = Release|Any CPU
+		{BA68D2AA-B0BE-4C8B-AD7C-B374A716865F}.Release|x64.Build.0 = Release|Any CPU
+		{BA68D2AA-B0BE-4C8B-AD7C-B374A716865F}.Release|x86.ActiveCfg = Release|Any CPU
+		{BA68D2AA-B0BE-4C8B-AD7C-B374A716865F}.Release|x86.Build.0 = Release|Any CPU
+		{4622F145-00FA-4BE3-939C-2CD67CAFC958}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4622F145-00FA-4BE3-939C-2CD67CAFC958}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4622F145-00FA-4BE3-939C-2CD67CAFC958}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{4622F145-00FA-4BE3-939C-2CD67CAFC958}.Debug|x64.Build.0 = Debug|Any CPU
+		{4622F145-00FA-4BE3-939C-2CD67CAFC958}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{4622F145-00FA-4BE3-939C-2CD67CAFC958}.Debug|x86.Build.0 = Debug|Any CPU
+		{4622F145-00FA-4BE3-939C-2CD67CAFC958}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4622F145-00FA-4BE3-939C-2CD67CAFC958}.Release|Any CPU.Build.0 = Release|Any CPU
+		{4622F145-00FA-4BE3-939C-2CD67CAFC958}.Release|x64.ActiveCfg = Release|Any CPU
+		{4622F145-00FA-4BE3-939C-2CD67CAFC958}.Release|x64.Build.0 = Release|Any CPU
+		{4622F145-00FA-4BE3-939C-2CD67CAFC958}.Release|x86.ActiveCfg = Release|Any CPU
+		{4622F145-00FA-4BE3-939C-2CD67CAFC958}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -52,5 +94,8 @@ Global
 	GlobalSection(NestedProjects) = preSolution
 		{4AE6501D-FCD7-4AC6-BFCA-13BD4E2B0482} = {827E0CD3-B72D-47B6-A68D-7590B98EB39B}
 		{A8EDFA8B-D4CB-4C33-90D6-40E45849C773} = {0AB3BF05-4346-4AA6-1389-037BE0695223}
+		{DB055AC4-1685-4A24-9779-7E1A37CA9986} = {827E0CD3-B72D-47B6-A68D-7590B98EB39B}
+		{BA68D2AA-B0BE-4C8B-AD7C-B374A716865F} = {827E0CD3-B72D-47B6-A68D-7590B98EB39B}
+		{4622F145-00FA-4BE3-939C-2CD67CAFC958} = {827E0CD3-B72D-47B6-A68D-7590B98EB39B}
 	EndGlobalSection
 EndGlobal
