+++ src/CtfDeck.Data/PersistenceModels/Sessions/SessionTarget.cs
@@ -1,0 +1,11 @@
+namespace CtfDeck.Data.PersistenceModels.Sessions;
+
+public class SessionTarget
+{
+    public Guid Id { get; set; }
+    public string Address { get; set; } = string.Empty;
+    public int? Port { get; set; }
+    public string Name { get; set; } = string.Empty;
+    public string Description { get; set; } = string.Empty;
+    public TargetType Type { get; set; }
+}
