+++ src/CtfDeck.Data/PersistenceModels/Sessions/TargetType.cs
@@ -1,0 +1,12 @@
+namespace CtfDeck.Data.PersistenceModels.Sessions;
+
+public enum TargetType
+{
+    Unknown = 0,
+    Web = 1,
+    Pwn = 2,
+    Crypto = 3,
+    Forensics = 4,
+    Reverse = 5,
+    Misc = 6
+}
