+++ src/CtfDeck.Data/Repositories/Scripts/ICustomScriptRepository.cs
@@ -1,0 +1,11 @@
+using CtfDeck.Terminal.Script.Models;
+
+namespace CtfDeck.Data.Repositories.Scripts;
+
+public interface ICustomScriptRepository
+{
+    CustomScript Create(string name, ScriptCategory category, string template);
+    List<CustomScript> GetAll();
+    bool Update(Guid id, string name, ScriptCategory category, string template);
+    bool Delete(Guid id);
+}
