+++ src/CtfDeck.ServerWs/CtfDeck.ServerWs.csproj
@@ -1,0 +1,16 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <ItemGroup>
+    <ProjectReference Include="..\CtfDeck.Terminal\CtfDeck.Terminal.csproj" />
+    <ProjectReference Include="..\CtfDeck.Data\CtfDeck.Data.csproj" />
+    <ProjectReference Include="..\CtfDeck.Contracts\CtfDeck.Contracts.csproj" />
+  </ItemGroup>
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>net8.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+  </PropertyGroup>
+
+</Project>
