--- src/CtfDeck.Terminal/Dto/CommandResponseDto.cs
@@ -1,6 +1,0 @@
-namespace CtfDeck.Terminal.Dto;
-
-public class CommandResponseDto
-{
-
-}
