--- src/CtfDeck.Terminal/Session/Data/SessionDbContext.cs
@@ -1,70 +1,0 @@
-using LiteDB;
-using SessionModel = CtfDeck.Terminal.Session.Models.Session;
-using CtfDeck.Terminal.Session.Models;
-using CtfDeck.Terminal.Script.Models;
-
-namespace CtfDeck.Terminal.Session.Data;
-
-public class SessionDbContext : IDisposable
-{
-    private static bool _mapperConfigured;
-    private static readonly object _mapperLock = new();
-
-    private readonly LiteDatabase _database;
-    private readonly ILiteCollection<SessionModel> _sessions;
-    private readonly ILiteCollection<CustomScript> _customScripts;
-    private bool _disposed;
-
-    public SessionDbContext(string? databasePath = null)
-    {
-        // Use in-memory database if path is ":memory:" or null for tests
-        var connectionString = databasePath == ":memory:"
-            ? "Filename=:memory:;Mode=Memory;Cache=Shared"
-            : $"Filename={databasePath ?? Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "ctfdeck_sessions.db")}";
-
-        ConfigureBsonMapper();
-        _database = new LiteDatabase(connectionString);
-        _sessions = _database.GetCollection<SessionModel>("sessions");
-        _sessions.EnsureIndex(x => x.Id, unique: true);
-        _sessions.EnsureIndex(x => x.Name);
-
-        _customScripts = _database.GetCollection<CustomScript>("customscripts");
-        _customScripts.EnsureIndex(x => x.Id, unique: true);
-    }
-
-    public ILiteCollection<SessionModel> Sessions => _sessions;
-    public ILiteCollection<CustomScript> CustomScripts => _customScripts;
-
-    private static void ConfigureBsonMapper()
-    {
-        if (_mapperConfigured) return;
-
-        lock (_mapperLock)
-        {
-            if (_mapperConfigured) return;
-
-            BsonMapper.Global.EnumAsInteger = true;
-
-            BsonMapper.Global.Entity<SessionModel>()
-                .Id(x => x.Id);
-
-            BsonMapper.Global.Entity<HistoryEntry>()
-                .Id(x => x.Id);
-
-            BsonMapper.Global.Entity<SessionTarget>()
-                .Id(x => x.Id);
-
-            BsonMapper.Global.Entity<CustomScript>()
-                .Id(x => x.Id);
-
-            _mapperConfigured = true;
-        }
-    }
-
-    public void Dispose()
-    {
-        if (_disposed) return;
-        _database.Dispose();
-        _disposed = true;
-    }
-}
