--- src/CtfDeck.Terminal/Session/Models/SessionTarget.cs
@@ -1,11 +1,0 @@
-namespace CtfDeck.Terminal.Session.Models;
-
-public class SessionTarget
-{
-    public Guid Id { get; set; }
-    public string Address { get; set; } = string.Empty;
-    public int? Port { get; set; }
-    public string Name { get; set; } = string.Empty;
-    public string Description { get; set; } = string.Empty;
-    public TargetType Type { get; set; }
-}
