--- src/CtfDeck.Terminal/Session/Models/TargetType.cs
@@ -1,12 +1,0 @@
-namespace CtfDeck.Terminal.Session.Models;
-
-public enum TargetType
-{
-    Unknown = 0,
-    Web = 1,
-    Pwn = 2,
-    Crypto = 3,
-    Forensics = 4,
-    Reverse = 5,
-    Misc = 6
-}
