+++ src/zzzzzzz
@@ -1,0 +1,98 @@
+.
+├── CtfDeck.Terminal
+│   ├── CtfDeck.Terminal.csproj
+│   ├── Dto
+│   │   └── CommandResponseDto.cs
+│   ├── Program.cs
+│   ├── Script
+│   │   ├── CustomScriptMessageHandler.cs
+│   │   ├── Models
+│   │   │   └── CustomScript.cs
+│   │   ├── Protocol
+│   │   │   ├── CustomScriptProtocolDeserializer.cs
+│   │   │   └── CustomScriptProtocolSerializer.cs
+│   │   ├── Repositories
+│   │   │   ├── CustomScriptRepository.cs
+│   │   │   └── ICustomScriptRepository.cs
+│   │   └── Services
+│   │       └── CustomScriptService.cs
+│   ├── Session
+│   │   ├── ActiveSessionManager.cs
+│   │   ├── Data
+│   │   │   └── SessionDbContext.cs
+│   │   ├── Models
+│   │   │   ├── HistoryEntry.cs
+│   │   │   ├── Session.cs
+│   │   │   ├── SessionMetadata.cs
+│   │   │   ├── SessionTarget.cs
+│   │   │   └── TargetType.cs
+│   │   ├── Protocol
+│   │   │   ├── SessionProtocolDeserializer.cs
+│   │   │   └── SessionProtocolSerializer.cs
+│   │   ├── Repositories
+│   │   │   ├── ISessionRepository.cs
+│   │   │   └── SessionRepository.cs
+│   │   ├── Services
+│   │   │   └── SessionService.cs
+│   │   └── SessionMessageHandler.cs
+│   ├── Terminal
+│   │   ├── Core
+│   │   │   ├── CommandResult.cs
+│   │   │   └── Common.cs
+│   │   ├── Execution
+│   │   │   ├── CommandPreprocessor.cs
+│   │   │   └── ProcessRunner.cs
+│   │   ├── Navigation
+│   │   │   └── DirectoryNavigator.cs
+│   │   ├── Shell
+│   │   │   ├── ShellDetector.cs
+│   │   │   └── ShellModels.cs
+│   │   ├── TerminalExecutor.cs
+│   │   └── TerminalService.cs
+│   ├── WebSocket
+│   │   ├── BinaryProtocol.cs
+│   │   ├── OutputBatcher.cs
+│   │   └── WebSocketServer.cs
+│   ├── bin
+│   │   └── Debug
+│   │       └── net8.0
+│   │           ├── CtfDeck.Terminal
+│   │           ├── CtfDeck.Terminal.deps.json
+│   │           ├── CtfDeck.Terminal.dll
+│   │           ├── CtfDeck.Terminal.pdb
+│   │           ├── CtfDeck.Terminal.runtimeconfig.json
+│   │           ├── LiteDB.dll
+│   │           └── ctfdeck_sessions.db
+│   └── obj
+│       ├── CtfDeck.Terminal.csproj.nuget.dgspec.json
+│       ├── CtfDeck.Terminal.csproj.nuget.g.props
+│       ├── CtfDeck.Terminal.csproj.nuget.g.targets
+│       ├── Debug
+│       │   └── net8.0
+│       │       ├── CtfDeck..BD72EF11.Up2Date
+│       │       ├── CtfDeck.Terminal.AssemblyInfo.cs
+│       │       ├── CtfDeck.Terminal.AssemblyInfoInputs.cache
+│       │       ├── CtfDeck.Terminal.GeneratedMSBuildEditorConfig.editorconfig
+│       │       ├── CtfDeck.Terminal.GlobalUsings.g.cs
+│       │       ├── CtfDeck.Terminal.assets.cache
+│       │       ├── CtfDeck.Terminal.csproj.AssemblyReference.cache
+│       │       ├── CtfDeck.Terminal.csproj.CoreCompileInputs.cache
+│       │       ├── CtfDeck.Terminal.csproj.FileListAbsolute.txt
+│       │       ├── CtfDeck.Terminal.dll
+│       │       ├── CtfDeck.Terminal.genruntimeconfig.cache
+│       │       ├── CtfDeck.Terminal.pdb
+│       │       ├── CtfDeck.Terminal.sourcelink.json
+│       │       ├── apphost
+│       │       ├── apphost.exe
+│       │       ├── ref
+│       │       │   └── CtfDeck.Terminal.dll
+│       │       └── refint
+│       │           └── CtfDeck.Terminal.dll
+│       ├── project.assets.json
+│       ├── project.nuget.cache
+│       ├── project.packagespec.json
+│       ├── rider.project.model.nuget.info
+│       └── rider.project.restore.info
+└── zzzzzzz
+
+28 directories, 68 files
